requirements.txt
python-telegram-bot>=21.0
yt-dlp
python-dotenv
gdown
requests
